+++ include/swift/Parse/Parser.h
@@ -1223,6 +1223,13 @@ public:
   ParserResult<TypeRepr> parseType(Diag<> MessageID,
                                    bool IsSILFuncDecl = false);
 
+  /// Parse a type optionally prefixed by a list of named opaque parameters. If
+  /// no params present, return 'type'. Otherwise, return 'type-named-opaque'.
+  ///
+  ///   type-named-opaque:
+  ///     generic-params type
+  ParserResult<TypeRepr> parseTypeWithOpaqueParams(Diag<> MessageID);
+
   ParserResult<TypeRepr>
     parseTypeSimpleOrComposition(Diag<> MessageID);
